Added support for makefiles
This uses a make command to output a dump of the make database, then parses out build targets from that output. Unfortunately there doesn't seem to be a good way to get make to output just the available build targets so some manual string manipulation is required.

Also refactored some of the config stuff so the alternate configuration parsers could be moved to their own directory.
